feat:保存发送记录

+++ model/sms_record.go
@@ -11,7 +11,7 @@ type SmsRecord struct {
 	SupplierName      string    `gorm:"column:supplier_name" db:"supplier_name" json:"supplier_name" form:"supplier_name"`                         //发送短信渠道商的名称
 	MsgContent        string    `gorm:"column:msg_content" db:"msg_content" json:"msg_content" form:"msg_content"`                                 //短信发送的内容
 	SeriesId          string    `gorm:"column:series_id" db:"series_id" json:"series_id" form:"series_id"`                                         //下发批次的ID
-	ChargingNum       int8      `gorm:"column:charging_num" db:"charging_num" json:"charging_num" form:"charging_num"`                             //计费条数
+	ChargingNum       uint64      `gorm:"column:charging_num" db:"charging_num" json:"charging_num" form:"charging_num"`                             //计费条数
 	ReportContent     string    `gorm:"column:report_content" db:"report_content" json:"report_content" form:"report_content"`                     //回执内容
 	Status            int8      `gorm:"column:status" db:"status" json:"status" form:"status"`                                                     //短信状态： 10.发送 20.成功 30.失败
 	SendDate          int       `gorm:"column:send_date" db:"send_date" json:"send_date" form:"send_date"`                                         //发送日期：20211112
@@ -19,6 +19,6 @@ type SmsRecord struct {
 	CreatedAt         time.Time `gorm:"column:created_at" db:"created_at" json:"created_at" form:"created_at"`                                     //创建时间
 }
 
-func (s *SmsRecord) TableName() string  {
+func (s *SmsRecord) TableName() string {
 	return "sms_record"
 }
